Move gitGetOutput and gitGetOutputCached to separate file
